uav-producer/requirements.txt
kafka-python
minio
rasterio
numpy
Pillow
opencv-python-headless
